refactor: organize methods in mock backend (collect all util functions in util.ts, collect all environment parsing sections in the main.ts)
